+++ src/runtime_src/xdp/profile/device/hal_device/xdp_hal_device.h
@@ -43,11 +43,11 @@ public:
   virtual int readTraceData(void* traceBuf, uint32_t traceBufSz, uint32_t numSamples, uint64_t ipBaseAddress, uint32_t& wordsPerSample);
 
   virtual size_t alloc(size_t sz, uint64_t memoryIndex);
-  virtual void free(uint32_t xdpBoHandle);
-  virtual void* map(uint32_t xdpBoHandle);
-  virtual void unmap(uint32_t xdpBoHandle);
-  virtual void sync(uint32_t xdpBoHandle, size_t sz, size_t offset, direction dir, bool async=false);
-  virtual uint64_t getDeviceAddr(uint32_t xdpBoHandle);
+  virtual void free(size_t xdpBoHandle);
+  virtual void* map(size_t xdpBoHandle);
+  virtual void unmap(size_t xdpBoHandle);
+  virtual void sync(size_t xdpBoHandle, size_t sz, size_t offset, direction dir, bool async=false);
+  virtual uint64_t getDeviceAddr(size_t xdpBoHandle);
   virtual void* getRawDevice() { return mHalDevice ; }
 };
 }
