+++ src/runtime_src/xdp/profile/device/xdp_base_device.h
@@ -43,12 +43,12 @@ public:
 
   // Only device RAM
   virtual size_t alloc(size_t sz, uint64_t memoryIndex) = 0;
-  virtual void free(uint32_t xdpBoHandle) = 0;
+  virtual void free(size_t xdpBoHandle) = 0;
 
-  virtual void* map(uint32_t xdpBoHandle) = 0;
-  virtual void unmap(uint32_t xdpBoHandle) = 0;
-  virtual void sync(uint32_t xdpBoHandle, size_t sz, size_t offset, direction dir, bool async=false) = 0;
-  virtual uint64_t getDeviceAddr(uint32_t xdpBoHandle) = 0;
+  virtual void* map(size_t xdpBoHandle) = 0;
+  virtual void unmap(size_t xdpBoHandle) = 0;
+  virtual void sync(size_t xdpBoHandle, size_t sz, size_t offset, direction dir, bool async=false) = 0;
+  virtual uint64_t getDeviceAddr(size_t xdpBoHandle) = 0;
 
   virtual double getDeviceClock() = 0;
   virtual uint64_t getTraceTime() = 0;
